Add fn. enough_resources_to_build and refactor can_build
Rename can_build to can_build_on_map. The new name specifies what the
function is actually computing (only if the proposed position on the
map is an allowable spot to build the building). Further, the new name
is better so as not to be confused with functions like
can_build_villager (which is defined for the TownCenter class).