Add empty classes: Army and Group
The Army class is intended to be able to easily select multiple military
units at once and give commands to them (like move or attack).

The Group class is intended for villagers. (For example, to have a
group of them build a building.)

I may end up making the group class general and have Army
inherit from Group.