Add method beginning_plus_the_rest to Vector class
The purpose of this method is to allow players to command units to move
more than 15 spots. The method then splits up a proposed movement
Vector into what will be done in the next turn and what will be done
afterwards.

It is necessary to also add a __add__ method to the Vector class (which
overrides the method defined for the Position class).