Write 2/3 of print_centered_at +more
Meaning of 2/3:
The function print_centered_at can now be used to print a portion of
the map. On both the left and right sides, the row number is also
printed. What is missing is printing the column numbers at the top and
the bottom.

more:
Fill out game_map some by adding stone, gold, food, and more wood.