Fix slight bug in fn print_centered_at
Before this fix, the function would never print the far right column nor
the bottom row. This fix changes that. (It also adds some Iron at the
right and bottom of the map.)

Also, the Bronze that was placed at regular intervals, proving the
printed indexes are correct, has been removed and replaced with other
bronze deposits.