Allow buildings with two word  names to by typed as such
This effects the following two functions:
 - extract_selected_obj
 - build_building

Before this change, I had already tried to make the function
build_building be able to handle two word building names. It did not
because the beginning of the function previously forced inpt_as_ls to
have length 4. Now it only forces it to have length at least 4.