Alter Vector method beginning_plus_the_rest
Change it to have an optional parameter distance_in_one_turn. When the
optional parameter is not there, this function should perform exactly as
it did before. I want to actually use the optional parameter when
building units (in the case that the build_position is greater than 6
spots away from the building's position).