Fix hidden "bug" on building buildings
Before this fix, when a villager is commanded to build a building, they
are never removed from player.commands['now']['build building']

To see that this fix works, add the following print statement to
main.py:
print('building commands:', player.commands['now']['build building'])

This bug was hidden because once a building is built, it doesn't really
matter that villagers continue to "build" it. (The Building method
build would just ignore the command.) Hence, the user would never
actually notice this bug.