Change Building.__init__ and use new player.building_position_pairs
In player.py, add the attribute building_position_pairs to the each
player's initialization. The purpose of this is to be able to access a
building given its location.

Whenever a new building is instantiated, they should be added to
player.building_position_pairs. Consequently, the function
Building.__init__ requires the new argument player. This forced an edit
for every place a building is instantiated.

The updated Building.__init__ function uses the new Building method
insert_into_building_position_pairs

Finally, since player.building_position_pairs is a dictionary with
Position instances as keys, Position instances need to be made hashable.
Hence the method __hash__ needs to be added to the Position class.
(Before this, Python gave an error that the Position class was not
hashable.)