Make load_game work (mostly)
The fn load_game ignores one small thing at this point:

If in a game, you command a building to research something while it is
already researching something else, and then if when prompted,
you change your mind and say "no" I don't actually want the same
building to add another research item to its research queue, then
if you save the game and
then load it, this assumed incorrectly that you DID want to add it
to the research queue. Making this simplification makes implementing
the fn load_game easier.