Add method compare_to
This is for testing purposes. It checks to see if two unit
instances are basically in the same state. (There is an error margin
for when comparing their positions
due to randomness in what the units do. Specifically,
villagers currently use randomness when choosing which resource
instance to collect from next.)