Fix bug and use a building alias for construction
When a building is under construction, the game uses a
BuildingUnderConstruction instance as a placeholder/alias to the
actual building until construction is complete. This fixes
the bug that different buildings of the same type built on the same
turn get the same number. Also, this allows the (alias) building to
be selected while it is being built. Hence, it also prints the
progress to completion of the construction.