Change game_map indexing
Now a Position(i, j) instance has i as its x coordinate and j as its
y coordinate. Similarly, when the map is printed centered at
(i, j), then i is the x coordinate and j is the y coordinate.

Also, the y indexing for the map now increases as you go up, instead
of what it was before: decreasing as you go up.