Change where unit costs are subtracted +more
Now, when a player commands that new units be built, the cost is
subtracted in the insert_commands.py module, as it is for building
buildings and researching things.

more:
Move where messages are added for newly created units. Move it to
the Unit class __init__ method so that it doesn't need to be repeated
for every individual type of unit.