Change and fill out Unit method can_move
Add GameMap method has_villager_at

One thing that is changed in the Unit.can_move method is that it no
longer checks that delta.magnitude <= 15. This change is necessary if
I plan on using it in implement_commands.py, where I am intending to
use FindPath and allow units to move to a position close to their
desired position, even if that position is more than 15 away (but
assuming the desired position is within 15). Note that the
insert_move_command function in insert_commands.py itself does check
that the position to move to is not too far away.